Add lane_to_marker and launch file.
Adds a Python node in styx_visualization that converts styx_msgs/Lane
to visualization_msgs/Marker.
Also adds a launch file and rviz configuration to display the markers
in rviz with the URDF model provided by Dataspeed.

To run:
roslaunch styx_visualization visualization.launch

The base waypoints are in green, the final waypoints are in blue, and
the traffic waypoints are in green.