Add "once" option to lane_to_marker
Because the base waypoints list is so large, it takes a lot of
processing to convert it to markers. Because it never changes,
there's no need to convert it more than once. The "once" option
tells the lane_to_marker node to unsubscribe from the topic after
converting it once. The latch option stores the generated markers
and publishes them to each new node who subscribes to the markers.