Efficiency improvements
This code was built/modified on top of older Udacity code; it
will need to be merged into the updated version.

server.py: send a dbw_enabled message when the connection is
first made, so that we know the initial status of dbw.
Efficiency fix: Experiments showed that on my machine, at least,
the simulator's throttle was sporadically going to zero (as shown
by "control" messages, as well as by fluctuating AccN values in the
simulator).  The fix is to send the simulator all 3 control
messages at once, with throttle as the third.

styx server.launch: Send output to screen.

dbw_node.py: Increase max_lat_accel into YawController so that
we can take curves better at high speed.

pure_pursuit_core.cpp: Turn a warning into an error message, so that
it is more visible.

waypoint_updater.py: At the end of the track, provide waypoints from
the beginning of the track to complete the loop.  Changed some
parameters to allow high-speed driving.