Use non-linear deceleration curve for braking.
    x = v^2 / (2*a)

Therefore:

    v = sqrt(x * 2 * a)

In this case, `x` is the distance from waypoint to stop waypoint and
`a` is the maximum rate of deceleration, which is set to 1.0 in the
`dbw_node`, so also set to 1.0 here.

Also added a temporary hack to subtract 4.0 meters from the stopping
distance to approximate stop lines, but this isn't a good long term
solution.