Changes throttle PID params and use only steering PID instead of yaw controller
I noticed that the yaw_controller does not work great if current
velocity and target velocity don't agree.

The throttle PID might also need more attention.  Right now it takes
the velocity difference (between current and target) as input.  It
might be better to take the needed acceleration??  It is still
unclear to me what a thorttle value of X meas for the accelleration
of the car