Publish distance to next light, even if not red
tl_detector.py: Modified value published in topic /traffic_waypoint.
If wp (the published value) is > 0, then wp is the waypoint of the
next red light.  If wp < 0, then abs(wp) is the waypoint of the next
light, but the next light is not red (or its state is unknown).

waypoint_updater.py: Modified value published in topic /tl_distance.
If tld (the published value) is > 0, then tld is the distance in
meters to the next light, and the next light is red.  If tld < 0,
then abs(tld) is the distance to the next light, and the next
light is not red (or its state is unknown).