Implement Waypoint Updater (Partial)
Implement partially the `waypoint_updater.py` node, following Udacity suggested
workflow.

The node takes the list of all waypoints and the current car position and publishes
the list of waypoints ahead of the car, sorted according to distance from the car.

Debugging is not easy until `dbw` node is also implemented. So I did my best to debug
singularly this node but I cannot guarantee this really works.