Added TFR Converter
The converter script can be executed from a directory. It will convert images from that directory (recursively) into TF Records - A training set and a testing set.

The TF Records are a binary format of the sample images, and can be used to fine tune TF models.

The script works, but several improvements could still be made:

* Allow the script to accept input arguments for input/output folders, and split ratio.
* Pass the label map by reference instead of hard-coding.
* Improve the shuffle-split algorithm, so that it maintains an even ratio of the classes.