Implemented rudimentary waypoint finding
The current implementation finds the waypoint closest to the car
based on euclidean distance in the XY plane and publishes LOOKAHEAD_WPS
many waypoints to the /final_waypoints topic.

It also performs a basic check that the next waypoint isn't behind the
car, assuming the car is facing the correct way and isn't wildly off
from the track.