Keep speed limit
Makes sure that the calculated velocities in waypoints provided to
the control subsystem is within the speed limit. Although this error
never should happen, this is not considered to be a fatal failure,
and it's handled by simply setting the speed within the limit and log
the error.

This change relies on all base_waypoints having the speed set to a
constant speed_limit when loaded. There was however some code
applying initial acceleration on the vehicle to the base_waypoints,
which makes no sense so it was removed. Such calculations will later
on be added on the final_waypoints. But for now, all
final_waypoints are set to the speed limit and the acceleration is
instead solely handled by the throttle PID controller.