Speed calculator
Helper class to calculate the speeds at given distance when
accelerating as much possible within given limitations to
acceleration and jerk.

Instead of solving the integral of the non-continuous third-degree
polynomial, an iterative solution is implemented where a look-up
table (LUT) containing speeds at discrete distances are calculated.
The speed at a specific distance is then approximated by linear
interpolation between the values in the LUT.