Optimal acceleration in final waypoints
Use the SpeedCalculator to find the target speed at each waypoint
when applying a limited amount of jerk and acceleration to reach the
max permitted velocity, e.g. the speed limit.

The velocity of the preceding waypoint is used as base point for the
speed calculation, instead of using the actual velocity from the
/current_velocity topic. This prevent problems related to delays and
errors in the control loop. The actual velocity is however used as
base point for the first calculation.