Add acceleration to Waypoint msg
Keep track of calculated acceleration by storing it in each waypoint,
which was convenient since we already store previously calculated
final_waypoints. The acceleration is for now only used within
waypoint_updater, but may later on also be useful for the control
modules, which is the more valid motivation for adding it to the
message.