Decelerate for traffic-lights using max jerk
Use the SpeedCalculator to find more physically correct velocities
when decelerating for traffic-lights. For now, max permitted
jerk and acceleration is always applied.

The STOPLINE_WPS_MARGIN is increased because current control modules
is not able to execute this deceleration. There are however other
issues to tune the controller, which might fix this problem so lets
keep the limits for now.