Only brake when possible to stop in time
If calculation shows that the vehicle is unable to decelerate in time
for a traffic light, then it will just continue without braking. This
is to handle the case when the traffic light change state from green
to yellow just as the car pass through.

The deceleration is also calculated with several jerk_limits, where a
lower limit is preferred when possible, still supporting limits up to
10 m/s^3 when needed.