Throttle PID coefficients
The Twiddle algorithm was running for about 24 hours and these
coefficients gave the best score.

The result from running a few tuning cycles with the same
coefficients does however actually give quite large variance to the
twiddle scores. I guess extraneous factors from running in the
simulator actually have larger impact on the twiddle score than the
changes made to the coefficients. So in order to improve the
coefficients further, a better measurement for the score is needed,
maybe by taking the average over several tuning cycles.