Stopline margin defined as a distance
Change the stopline margin from a number of waypoints into a
distance. The motivation for this is that the distance between
waypoints varies, so the number of waypoints to use as margin would
then need to be individually specified for each TL. With this change,
the vehicle will still try to stop at a specific waypoint, but this
waypoint is now calculated to be at least at the specified distance
behind the stopline.