Throttle tuning bugfix
In the tuning process, a negative value for the integral coefficient
was chosen. This gave better result in the ~2 min tuning cycles, but
then the accumulated error was reset before the next tuning cycle.
However, when running the vehicle for longer distances, this error
continue to increase since the tuning is made to prevent ever
exceeding the target velocity. Eventually, this negative coefficient
will make the vehicle go slower and slower.

As a temporary fix the integral coefficient is set to 0.