CMakeLists.txt produces error - marked content to be incorrect
test: CMakeLists.txt produces error - marked content to be incorrect

The original CMarkeLists.txt file contained purely the name of the file. Symbolic links don't work on GitHub. Therefore, the path to the ros directory's CMakeLists.txt file was added. But this doesn't work in the docker environment.

Therefore, the CMakeLists.txt has now a comment added describing the problem. This doesn't solve the issue though. Further investigation and test is required.