Update lowpass.py
It is a minor change but it is very clear now what this class is going to do. 
i.e 0<a<1 is a weight multiplied to current measurement and (1-a) to the last value function of last_measurement. 

I would go ahead and remove b altogether and just use a and (1-a)